requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
PyMySQL==1.1.1
bcrypt==4.2.0
pycryptodome==3.20.0
requests==2.32.3
pytest==8.3.4
Pillow>=10.0.0
pyotp
qrcode

